left with create, update and post design

+++ include/postDB.php
@@ -1,42 +1,58 @@
-<?php    
+<?php 
+        session_start();
+        global $uid;
+        if (isset($_SESSION["user_id"]))
+        {
+            $uid = $_SESSION["user_id"];
+        }
         $action = $_POST['action'];
         $pid = $_POST['id'];
 
         if ($action == "add_fav_post")
         {
-            add_fav_post($pid);
+            add_fav_post($pid, $uid);
         }else if ($action == "remove_fav_post")
         {
-            remove_fav_post($pid);
+            if (is_array($pid))
+            {
+                foreach ($pid as $i)
+                {
+                    remove_fav_post($i, $uid);
+                }
+            }else
+            {
+                remove_fav_post($pid, $uid);
+            }
+            
         }
 ?>
 <?php
-session_start();
-function get_user_id()
-{
-    global $user_id;
 
-    $config = parse_ini_file('../../private/db-config.ini');
-    $conn = new mysqli($config['servername'], $config['username'], $config['password'], $config['dbname']);
+// function get_user_id()
+// {
+//     global $user_id;
 
-    if ($conn->connect_error) {
-        $errorMsg = "Connection failed: " . $conn->connect_error;
-        $success = false;
-    }else
-    {
-        $stmt = $conn->prepare("SELECT user_id FROM user WHERE email=?");
-        $stmt->bind_param("s", $email);
+//     $config = parse_ini_file('../../private/db-config.ini');
+//     $conn = new mysqli($config['servername'], $config['username'], $config['password'], $config['dbname']);
 
-        $stmt->execute();
-        $result = $stmt->get_result();
-        $value = $result->fetch_object();
-        $user_id = $value->user_id;
-        $stmt->close();
-    }
-    $conn->close();
+//     if ($conn->connect_error) {
+//         $errorMsg = "Connection failed: " . $conn->connect_error;
+//         $success = false;
+//     }else
+//     {
+//         $stmt = $conn->prepare("SELECT user_id FROM user WHERE email=?");
+//         $stmt->bind_param("s", $email);
+
+//         $stmt->execute();
+//         $result = $stmt->get_result();
+//         $value = $result->fetch_object();
+//         $user_id = $value->user_id;
+//         $stmt->close();
+//     }
+//     $conn->close();
 
 
-}
+// }
 
 //Retrieve all the post
 function display()
@@ -111,7 +127,7 @@ function display()
     //Retrieve function to display post
     function display_post()
     {
-        global $s;
+        global $uid;
         $postid = $_GET["id"];
         $config = parse_ini_file('../../private/db-config.ini');
         $conn = new mysqli($config['servername'], $config['username'], $config['password'], $config['dbname']);
@@ -169,12 +185,16 @@ function display()
                         echo '</div>';
                         echo '<div class="col-md-3 col-sm-6 middle">';
                        
-                            
+                         if (isset($_SESSION['email'])){   
                             echo '<div class="edit_delete">';
+                            if ($uid == $row["user_id"])
+                            {
                                 echo '<ul>';
                                     echo '<li class="pl"><a href="deletePost.php?id='.$postid.'">Delete<a>';
                                     echo '<li class="pl"><a href="updatePost.php?id='.$postid.'">Edit</a>';
                                 echo '</ul>';
+                            }else
+                            {
                                 echo '<i id="unbookmark2" data-id="'.$row["post_id"].'" class="material-icons fr">bookmark</i>';
                                 echo '<i id="bookmark2"  data-id="'.$row["post_id"].'" class="material-icons fr">bookmark_border</i>';
                                 if(checkFav())
@@ -184,9 +204,10 @@ function display()
                                 {
                                     echo '<i id="bookmark" data-id="'.$row["post_id"].'" class="material-icons fr">bookmark_border</i>';
                                 }
+                            }
                             echo '</div>';                              
                        
-                        
+                        }
                         echo '</div>';
                     echo '</div>';
                     echo '<div class="post-img">';
@@ -295,10 +316,10 @@ function display()
 
     
     // Insert bookmarked post
-    function add_fav_post($pid)
+    function add_fav_post($pid, $uid)
     {
         $postid = $pid;
-        $uid = "2";
+        //$uid = "1";
 
          $config = parse_ini_file($_SERVER["DOCUMENT_ROOT"].'/../private/db-config.ini');    
          $conn = new mysqli($config['servername'], $config['username'], $config['password'], $config['dbname']);
@@ -310,7 +331,6 @@ function display()
             echo $errorMsg;
         }else
         {
-            echo $GLOBALS['s'];
             $stmt = $conn->prepare("INSERT INTO favourite (post_id, user_id) VALUES(?,?)");
             
 
@@ -328,11 +348,10 @@ function display()
 
     
     //Remove favourite post
-    function remove_fav_post($pid)
+    function remove_fav_post($pid, $uid)
     {
         global $x, $servername;
         $postid = $pid;
-        $uid = "1";
 
         $config = parse_ini_file($_SERVER["DOCUMENT_ROOT"].'/../private/db-config.ini');
         $conn = new mysqli($config['servername'], $config['username'], $config['password'], $config['dbname']);
@@ -362,8 +381,11 @@ function display()
     //Check if the user has fav the post
     function checkFav()
     {
+        if (isset($_SESSION["user_id"]))
+        {
+            $uid = $_SESSION["user_id"];
+        }
         $postid = $_GET["id"];
-        $uid = "1";
         $config = parse_ini_file('../../private/db-config.ini');
         $conn = new mysqli($config['servername'], $config['username'], $config['password'], $config['dbname']);
 
@@ -394,7 +416,11 @@ function display()
     function display_fav_post()
     {
         global $fav_page_no, $fav_total_no_ofpages;
-        $uid = "1";
+
+        if (isset($_SESSION["user_id"]))
+        {
+            $uid = $_SESSION["user_id"];
+        }
 
         if($_GET["page"] != "")
         {
@@ -443,8 +469,11 @@ function display()
                 
                     $date = date("jS M Y",strtotime($row["p_publish_date"]));
                     $content = substr($row["p_content"],0,100);
-                    echo '<div class="col-md-4 col-sm-6 click shake" data-id='.$row["f_post_id"].'>';
-                    echo '<div class="blog post effect">';
+                    echo '<div class="col-md-4 col-sm-6">';
+                    echo '<span class="blog post effect cross">';
+                    echo '<input class="checkmate" type="checkbox" value='.$row["f_post_id"].' name="fav">';
+                    echo '</span>';
+                    echo '<div class="blog post effect click" data-id='.$row["f_post_id"].'>';
                     echo '<div class="blog-image">';
                     echo '<img src="image/'.$row["picture_image"].'" alt="'.$row["picture_image"].'">';
                     echo '</div>';
@@ -467,6 +496,93 @@ function display()
 
         }
 
+    }
+
+    //display post by user
+    function display_mypost()
+    {   
+        if (isset($_SESSION["user_id"]))
+        {
+            $uid = $_SESSION["user_id"];
+        }
+
+        $uid = "5";
+
+        global $page_no, $total_no_ofpages;
+
+        if($_GET["page"] != "")
+        {
+            $page_no = $_GET["page"];
+        }
+        $max_records_per_page = 9;
+        $offset = ($page_no - 1)*$max_records_per_page;
+        $previous_page = $page_no - 1;
+        $next_page = $page_no + 1;
+    
+    
+        $config = parse_ini_file('../../private/db-config.ini');
+        $conn = new mysqli($config['servername'], $config['username'], $config['password'], $config['dbname']);
+        if ($conn->connect_error) {
+            $errorMsg = "Connection failed: " . $conn->connect_error;
+            $success = false;
+        } else {
+            $stmt = $conn->prepare("SELECT * FROM post WHERE user_id=?");
+
+            $stmt->bind_param("i", $uid);
+            $stmt->execute();
+            $result_records = $stmt->get_result();
+            $total_post = $result_records->num_rows;
+            $total_no_ofpages = ceil($result_records->num_rows/$max_records_per_page);
+    
+    
+            $stmt = $conn->prepare("SELECT  p.post_id AS p_post_id, p.user_id AS p_user_id, p.author AS p_author,
+            p.title AS p_title, p.content AS p_content, p.publish_date AS p_publish_date,
+            picture.image AS picture_image FROM post p LEFT JOIN image AS picture
+            ON picture.post_id = p.post_id WHERE p.user_id=?
+            ORDER BY p.post_id DESC LIMIT " . $offset . "," . $max_records_per_page);
+
+            $stmt->bind_param("i", $uid);
+            $stmt->execute();
+            $result = $stmt->get_result();
+            if ($result->num_rows > 0) {
+                while ($row = mysqli_fetch_array($result)) {
+    
+                    if ($row["picture_image"] == NULL)
+                    {
+                        $row["picture_image"] = "defaultimg.png";
+                    }
+                    
+                    $date = date("jS M Y",strtotime($row["p_publish_date"]));
+                    $content = substr($row["p_content"],0,100);
+                    echo '<div class="col-md-4 col-sm-6 " >';
+                    echo '<span class="blog post effect del">';
+                    echo '<input class="checkmate" type="checkbox" value='.$row["p_post_id"].' name="delete">';
+                    echo '</span>';
+                    echo '<span class="blog post effect cross">';
+                    echo '<input class="radio" type="radio" value='.$row["p_post_id"].' name="edit">';
+                    echo '</span>';
+                    echo '<div class="blog post effect click" data-id='.$row["p_post_id"].'>';
+                    echo '<div class="blog-image">';
+                    echo '<img src="image/'.$row["picture_image"].'" alt="'.$row["picture_image"].'">';
+                    echo '</div>';
+                    echo '<div class="blog-title">';
+                    echo '<h3>' . $row["p_title"] . '</h3>';
+                    echo '</div>';
+                    echo '<div class="blog-content">' . $content . ' ... 
+                    <a href="viewpost.php?id=' .$row["p_post_id"].'">Read more</a></div>';
+                    echo '<div class="blog-footer">';
+                    echo '<ul class="post-info">';
+                    echo '<li><i class="material-icons edit">date_range</i>' . $date . '</li>';
+                    echo '<li><i class="material-icons edit">create</i>' . $row["p_author"] . '</li>';
+                    echo '</ul>';
+                    echo '</div>';
+                    echo '</div>';
+                    echo '</div>';
+                }
+                $stmt->close();
+            }
+        }
+        $conn->close();
     }
 
     //Count number of comments
